first pass at making editors cross-platform
