+++ app/src/lib/app-state.ts
@@ -19,7 +19,7 @@ import { IMenu } from '../models/app-menu'
 import { IRemote } from '../models/remote'
 import { WindowState } from './window-state'
 import { RetryAction } from './retry-actions'
-import { ExternalEditor } from '../models/editors'
+import { ExternalEditor } from '../lib/editors'
 import { PreferencesTab } from '../models/preferences'
 import { Shell } from './shells'
 import { CloneRepositoryTab } from '../models/clone-repository-tab'
