+++ app/src/lib/editors/darwin.ts
@@ -1,8 +1,28 @@
 import * as Path from 'path'
-import { ExternalEditor } from '../../models/editors'
 import { pathExists } from '../file-system'
 import { LookupResult, FoundEditor } from './shared'
 import { assertNever } from '../fatal-error'
+
+export enum ExternalEditor {
+  Atom = 'Atom',
+  VisualStudioCode = 'Visual Studio Code',
+  SublimeText = 'Sublime Text',
+}
+
+export function parse(label: string): ExternalEditor | null {
+  if (label === ExternalEditor.Atom) {
+    return ExternalEditor.Atom
+  }
+
+  if (label === ExternalEditor.VisualStudioCode) {
+    return ExternalEditor.VisualStudioCode
+  }
+  if (label === ExternalEditor.SublimeText) {
+    return ExternalEditor.SublimeText
+  }
+
+  return null
+}
 
 /**
  * appPath will raise an error if it cannot find the program.
