+++ app/src/lib/editors/index.ts
@@ -1,2 +1,3 @@
 export * from './lookup'
 export * from './launch'
+export { ExternalEditor, parse } from './shared'
