+++ app/src/lib/stores/app-store.ts
@@ -37,11 +37,7 @@ import { Branch, BranchType } from '../../models/branch'
 import { TipState } from '../../models/tip'
 import { CloningRepository } from '../../models/cloning-repository'
 import { Commit } from '../../models/commit'
-import {
-  ExternalEditor,
-  parse as parseExternalEditor,
-} from '../../models/editors'
-import { getAvailableEditors } from '../editors'
+import { ExternalEditor, getAvailableEditors, parse } from '../editors'
 import { CloningRepositoriesStore } from './cloning-repositories-store'
 import { IGitHubUser } from '../databases/github-user-database'
 import { GitHubUserStore } from './github-user-store'
@@ -938,7 +934,7 @@ export class AppStore {
   private async getSelectedExternalEditor(): Promise<ExternalEditor | null> {
     const externalEditorValue = localStorage.getItem(externalEditorKey)
     if (externalEditorValue) {
-      const value = parseExternalEditor(externalEditorValue)
+      const value = parse(externalEditorValue)
       if (value) {
         return value
       }
