+++ app/src/ui/preferences/advanced.tsx
@@ -5,7 +5,7 @@ import { LinkButton } from '../lib/link-button'
 import { Row } from '../../ui/lib/row'
 import { SamplesURL } from '../../lib/stats'
 import { Select } from '../lib/select'
-import { ExternalEditor, parse as parseEditor } from '../../models/editors'
+import { ExternalEditor, parse as parseEditor } from '../../lib/editors'
 import { Shell, parse as parseShell } from '../../lib/shells'
 import { TextBox } from '../lib/text-box'
 import { enablePreviewFeatures } from '../../lib/feature-flag'
